# public/docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Wudysoft API",
    "description": "Wudysoft API provides powerful RESTful AI services to integrate with your applications.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://wudynext.netlify.app",
      "description": "Original"
    },
    {
      "url": "https://wudysoft.us.kg",
      "description": "Alternative"
    }
  ],
  "paths": {
...
    "/api/random/tools": {
      "get": {
        "summary": " Api Random Tools",
        "description": "Endpoint ini untuk mengambil data  Api Random Tools.",
        "operationId": "_api_random_tools",
        "tags": ["RANDOM"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Permintaan berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string", "example": "Sukses"}
                  }
                }
              }
            }
          },
          "400": {"description": "Permintaan tidak valid"},
          "500": {"description": "Kesalahan server"}
        }
      }
    },
...
    "/api/tools/shortlink": {
      "get": {
        "summary": " Api Tools Shortlink",
        "description": "Endpoint ini untuk mengambil data  Api Tools Shortlink.",
        "operationId": "_api_tools_shortlink",
        "tags": ["TOOLS"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Permintaan berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {"type": "string", "example": "Sukses"}
                  }
                }
              }
            }
          },
          "400": {"description": "Permintaan tidak valid"},
          "500": {"description": "Kesalahan server"}
        }
      }
    },
...
  }
}